Added compatibility with Linux 5.3
Linux 5.3 made some changes to task_struct following the removal of
the tsk_nr_cpus_allowed wrapper, which added a pointer
to task_struct pointing to the cpumask_t cpus_allowed.
For some reason, this included a renaming of cpus_allowed to cpus_mask (see
https://lore.kernel.org/lkml/20190423142636.14347-1-bigeasy@linutronix.de/),
which breaks x86_a's compatibility with Linux 5.3 and newer.
This commit fixes that.

Changes:

* Add conditional usage of current->cpu_ptr instead of &(current->cpus_allowed)
if Kernel is 5.3 or newer