require CONFIG_THREAD_INFO_IN_TASK=y
Previous kernel version provided the current CPU inside of task_struct
directly.
This is not the case anymore, and the current CPU has to be extracted by
using information from thread_info, which is only provided if
CONFIG_THREAD_INFO_IN_TASK is enabled in the kernel config.

Builds without CONFIG_THREAD_INFO_IN_TASK will be aborted.

Note: I am not confident that this is the complete and correct solution,
however the most practical one. (I.e. feel free to overrule it down the
road)